refactor(broker): group cross-service infrastructure under shared/
Moves task.rs (RetryTask, BrokerService, Supervisor), errors.rs
(CodedError, coded_error_impl!, BrokerFailure), config.rs (Config,
ConfigLock, ConfigWatcher), and prioritization.rs (the unified
priority sorting core used by both order_pricer and order_committer)
into a new shared/ subdirectory.

shared/mod.rs is a thin barrel that declares the four submodules.

lib.rs replaces the four individual `mod` declarations with one
`pub(crate) mod shared;` and three re-exports that preserve every
existing import path:

    pub use shared::config;                         // was pub mod config
    pub(crate) use shared::{errors, prioritization, task};

No source changes outside lib.rs and the new mod.rs are needed —
existing `crate::config::X`, `crate::errors::X`, `crate::task::X`,
`crate::prioritization::X` imports across the crate keep resolving
through the re-exports. Pure file relocation.
