+++ crates/broker/src/lib.rs
@@ -31,10 +31,8 @@ mod broker;
 pub(crate) mod chain_monitor;
 pub(crate) mod chain_monitor_v2;
 pub(crate) mod channels;
-pub mod config;
 mod db;
 pub use db::{broker_sqlite_url_for_chain, DbObj, SqliteDb};
-pub(crate) mod errors;
 pub mod futures_retry;
 pub(crate) mod market_monitor;
 pub(crate) mod offchain_market_monitor;
@@ -44,7 +42,6 @@ pub(crate) mod order_locker;
 pub(crate) mod order_pricer;
 mod order_types;
 mod price_oracle;
-pub(crate) mod prioritization;
 pub mod provers;
 pub(crate) mod proving;
 pub(crate) mod reaper;
@@ -53,9 +50,11 @@ pub(crate) mod rpc_retry_policy;
 pub mod rpcmetrics;
 pub mod sequential_fallback;
 pub(crate) mod service_runner;
+pub(crate) mod shared;
+pub use shared::config;
+pub(crate) use shared::{errors, prioritization, task};
 pub(crate) mod storage;
 pub(crate) mod submitter;
-pub(crate) mod task;
 pub(crate) mod telemetry;
 pub mod utils;
 pub mod version_check;
