+++ crates/broker/src/shared/errors.rs
