+++ crates/broker/src/shared/mod.rs
@@ -1,0 +1,22 @@
+// Copyright 2026 Boundless Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Cross-service shared infrastructure: error coding, supervisor / retry task
+//! plumbing, configuration plumbing, and the prioritization core used by
+//! both the order pricer and the order committer.
+
+pub mod config;
+pub mod errors;
+pub mod prioritization;
+pub mod task;
